# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "DesiredNodeInput": {
        "additionalProperties": false,
        "properties": {
          "discord_id": {
            "anyOf": [
              {
                "pattern": "^[1-9][0-9]{0,19}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discord Id"
          },
          "logical_key": {
            "maxLength": 256,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Logical Key",
            "type": "string"
          },
          "presence": {
            "$ref": "#/components/schemas/NodePresence",
            "default": "PRESENT"
          },
          "properties": {
            "additionalProperties": true,
            "title": "Properties",
            "type": "object"
          },
          "relations": {
            "items": {
              "$ref": "#/components/schemas/RelationInput"
            },
            "maxItems": 3,
            "title": "Relations",
            "type": "array"
          },
          "resource_type": {
            "$ref": "#/components/schemas/ResourceType"
          },
          "symbol": {
            "anyOf": [
              {
                "maxLength": 256,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9_.:-]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          }
        },
        "required": [
          "logical_key",
          "resource_type"
        ],
        "title": "DesiredNodeInput",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NodePresence": {
        "enum": [
          "PRESENT",
          "ABSENT"
        ],
        "title": "NodePresence",
        "type": "string"
      },
...
      "PlanCreate": {
        "additionalProperties": false,
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/DesiredNodeInput"
            },
            "maxItems": 500,
            "minItems": 1,
            "title": "Nodes",
            "type": "array"
          },
          "schema_version": {
            "const": "did-dsg-v1",
            "default": "did-dsg-v1",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "nodes"
        ],
        "title": "PlanCreate",
        "type": "object"
      },
...
      "ReferenceKind": {
        "enum": [
          "LOGICAL",
          "DISCORD_ID",
          "SYMBOL"
        ],
        "title": "ReferenceKind",
        "type": "string"
      },
      "RelationInput": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/ReferenceKind"
          },
          "name": {
            "enum": [
              "parent",
              "channel",
              "subject"
            ],
            "title": "Name",
            "type": "string"
          },
          "value": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "name",
          "kind",
          "value"
        ],
        "title": "RelationInput",
        "type": "object"
      },
      "ResourceType": {
        "enum": [
          "GUILD",
          "ROLE",
          "CATEGORY",
          "CHANNEL",
          "OVERWRITE"
        ],
        "title": "ResourceType",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Discord Infrastructure Designer",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/guilds/{guild_id}/audit": {
      "get": {
        "operationId": "list_audit_api_v1_guilds__guild_id__audit_get",
        "parameters": [
          {
            "in": "path",
            "name": "guild_id",
            "required": true,
            "schema": {
              "title": "Guild Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Audit Api V1 Guilds  Guild Id  Audit Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Audit",
        "tags": [
          "stage-07-dashboard"
        ]
      }
    },
...
    "/api/v1/guilds/{guild_id}/plans": {
      "get": {
        "operationId": "list_plans_api_v1_guilds__guild_id__plans_get",
        "parameters": [
          {
            "in": "path",
            "name": "guild_id",
            "required": true,
            "schema": {
              "title": "Guild Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Plans Api V1 Guilds  Guild Id  Plans Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Plans",
        "tags": [
          "stage-07-dashboard"
        ]
      },
      "post": {
        "operationId": "create_plan_api_v1_guilds__guild_id__plans_post",
        "parameters": [
          {
            "in": "path",
            "name": "guild_id",
            "required": true,
            "schema": {
              "title": "Guild Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 160,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Plan Api V1 Guilds  Guild Id  Plans Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Plan",
        "tags": [
          "stage-05-plans"
        ]
      }
    },
...
    "/api/v1/ui/locales/{locale}/catalog/{catalog_version}": {
      "get": {
        "operationId": "ui_locale_catalog_api_v1_ui_locales__locale__catalog__catalog_version__get",
        "parameters": [
          {
            "in": "path",
            "name": "locale",
            "required": true,
            "schema": {
              "title": "Locale",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "catalog_version",
            "required": true,
            "schema": {
              "title": "Catalog Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ui Locale Catalog",
        "tags": [
          "stage-07-dashboard"
        ]
      }
    },
    "/auth/discord/callback": {
      "get": {
        "operationId": "discord_callback_auth_discord_callback_get",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "error",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Discord Callback",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/discord/login": {
      "get": {
        "operationId": "discord_login_auth_discord_login_get",
        "parameters": [
          {
            "in": "query",
            "name": "return_to",
            "required": false,
            "schema": {
              "default": "/",
              "title": "Return To",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Discord Login",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}